refactor: folders name and style const names
