+++ src/components/index.js
@@ -1,4 +1,4 @@
-import Post from './post/Post';
-import Fab from './fab/Fab';
+import Post from './Post/Post';
+import Fab from './Fab/Fab';
 
 export {Post, Fab};
