+++ src/screens/feed/Feed.styles.js
@@ -1,6 +1,6 @@
 import {StyleSheet} from 'react-native';
 
-export const TimelineStyles = StyleSheet.create({
+export const styles = StyleSheet.create({
   container: {
     flex: 1,
   },
